enhance(archive): change module name
